+++ src/clj/py99/routes/home.clj
@@ -173,7 +173,6 @@
 
 ;; --------------------
 ;; validations
-;; use declare? no.
 
 (defn- make-tempfile [dir login suffix]
   (str dir (System/nanoTime) "-" login suffix))
@@ -303,8 +302,6 @@
     (when-not (every? true?  (map starts-with-def-import-from-indent? lines))
       (throw (Exception. "回答中に実行文があるのはまずい。")))))
 
-
-
 (defn- ruff-path
   "ruff version 0.8.0 is out. 2024-11-23"
   []
@@ -379,15 +376,17 @@
 (defn- validate
   "Return nil if all validations success, or raize exeption."
   [num answer login]
-  (let [stripped (strip answer)]
+  (let [stripped (strip answer)
+        expanded-answer (expand-includes answer login)]
     (try
       (not-empty-test stripped)
-      (not-same-md5-login stripped login)
+      (when-not (u/dev?)
+        (not-same-md5-login stripped login))
       (has-docstring-test answer)
       (no-exec-statements answer)
       (ruff-formatter (str/trim (remove-comments answer)) login)
-      (doctest-test answer login)
-      (pytest-test num (expand-includes answer login))
+      (doctest-test expanded-answer login)
+      (pytest-test num expanded-answer)
       nil
       (catch Exception e
         (log/info "exception" (.getMessage e))
@@ -571,7 +570,6 @@
 
 (defn rank-submissions [request]
   (let [login (login request)]
-    ;; (log/debug "rank-submissions" login)
     (layout/render request "ranking-all.html"
                    {:data (db/submissions)
                     :title "Ranking Submissions"
@@ -580,7 +578,6 @@
 
 (defn rank-solved [request]
   (let [login (login request)]
-    ;; (log/debug "rank-solved" login)
     (layout/render request "ranking-all.html"
                    {:data (db/solved)
                     :title "Ranking Solved"
@@ -592,7 +589,6 @@
         data (map (fn [x] {:login (:from_login x)
                            :count (:count x)})
                   (db/comments-counts))]
-    ;; (log/debug "rank-comments" login)
     (layout/render request "ranking-all.html"
                    {:data data
                     :title "Comments Ranking"
@@ -601,7 +597,6 @@
 
 (defn answers-by-problems [request]
   (let [data (db/answers-by-problems)]
-    ;; (log/debug "answers-by-problems" (login request))
     (layout/render request "answers-by-problems.html"
                    {:data (reverse data)
                     :title "Answers by Problems"
@@ -627,13 +622,11 @@
 
 (defn list-stocks [request]
   (let [login (login request)]
-    ;; (log/debug "list-stocks" login)
     (layout/render request "stocks.html"
                    {:stocks (db/stocks? {:login login})})))
 
 (defn list-todays
   [{{:keys [date]} :path-params :as request}]
-  ;; (log/debug "list-todays" date)
   (if (re-matches #"\d\d\d\d-\d\d-\d\d" date)
     (layout/render request "todays.html"
                    {:date date
@@ -697,7 +690,6 @@
   (let [login (or (get-in request [:path-params :login])
                   (get-in request [:params :login]))
         user (get-user login)]
-    ;; (log/info "user-class" "login" login "user" user)
     (layout/render request "user-class.html" {:user user})))
 
 (defn home-routes []